Dashboard/requirements.txt
pandas==2.2.3
plotly==5.24.1
python_bcrypt==0.3.2
streamlit==1.40.2
streamlit_authenticator==0.4.1
